refactor: use fillStyle, stroke method
# Conflicts:
#	apps/chart/src/helpers/style.ts
